Reuse open_dataset logic in generate.py + rename minio file to avoid name collision during import.
